DONE BlockRoomMembers

+++ syemessenger/src/main/java/io/syemessenger/api/room/RoomService.java
@@ -3,10 +3,10 @@ package io.syemessenger.api.room;
 import static io.syemessenger.api.Pageables.toPageable;
 
 import io.syemessenger.api.ServiceMessage;
-import io.syemessenger.api.account.AccountInfo;
 import io.syemessenger.api.account.AccountMappers;
-import io.syemessenger.api.account.repository.Account;
 import io.syemessenger.api.account.repository.AccountRepository;
+import io.syemessenger.api.room.repository.BlockedMember;
+import io.syemessenger.api.room.repository.BlockedRepository;
 import io.syemessenger.api.room.repository.Room;
 import io.syemessenger.api.room.repository.RoomRepository;
 import io.syemessenger.websocket.SessionContext;
@@ -14,8 +14,8 @@ import jakarta.inject.Named;
 import java.time.Clock;
 import java.time.LocalDateTime;
 import java.time.temporal.ChronoUnit;
+import java.util.ArrayList;
 import java.util.List;
-import java.util.Optional;
 import org.springframework.dao.DataAccessException;
 import org.springframework.data.domain.Page;
 
@@ -24,10 +24,15 @@ public class RoomService {
 
   private final RoomRepository roomRepository;
   private final AccountRepository accountRepository;
+  private final BlockedRepository blockedRepository;
 
-  public RoomService(RoomRepository roomRepository, AccountRepository accountRepository) {
+  public RoomService(
+      RoomRepository roomRepository,
+      AccountRepository accountRepository,
+      BlockedRepository blockedRepository) {
     this.roomRepository = roomRepository;
     this.accountRepository = accountRepository;
+    this.blockedRepository = blockedRepository;
   }
 
   public void createRoom(SessionContext sessionContext, CreateRoomRequest request) {
@@ -290,11 +295,29 @@ public class RoomService {
       return;
     }
 
+    if (!sessionContext.accountId().equals(room.owner().id())) {
+      sessionContext.sendError(403, "Not room owner");
+      return;
+    }
+
+    if (memberIds.contains(room.owner().id())) {
+      sessionContext.sendError(400, "Cannot block room owner");
+      return;
+    }
+
+    final var blockedMembers = new ArrayList<BlockedMember>();
+    for (var memberId : memberIds) {
+      blockedMembers.add(new BlockedMember().roomId(roomId).accountId(memberId));
+    }
+
     try {
       roomRepository.deleteRoomMembers(roomId, memberIds);
-      roomRepository.blockMembers(roomId, memberIds);
+      blockedRepository.saveAll(blockedMembers);
       sessionContext.send(new ServiceMessage().qualifier("blockRoomMembers").data(roomId));
     } catch (Exception e) {
+      if (e.getMessage().contains("is not present in table")) {
+        sessionContext.sendError(404, "Account not found");
+      }
       sessionContext.sendError(500, e.getMessage());
     }
   }
